test(engine): split poll_test.go to mirror poll.go's file split
Move TestRunProbeAndDeepFetch_*/TestColdStart_*/TestWebhookModeStartup_*/
TestProbeNewItem_* into the existing terminal_test.go; move
TestRunStartupTransientLabelScan_*/TestCheckAllowAutoMerge_* into the
existing startup_test.go; move TestLayer1StatusRefresh_*/
TestReconcileLoop_RunsWithoutWebhookManager into a new reconcile_test.go.
poll_test.go retains TestPoll_*/TestCatchupLoop_*/TestYoloCatchup*/
TestCruiseCatchUp*/TestHandleMergeTrainBatch_*/TestGroupQueuedByRepo* and
worker-lifecycle tests, per ADR-010's test-mirrors-implementation
convention.
