+++ engine/poll_test.go
@@ -2,7 +2,6 @@ package engine
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -18,7 +17,6 @@ import (
 	"github.com/handarbeit/fabrik/internal/itemstate"
 	"github.com/handarbeit/fabrik/stages"
 	"github.com/handarbeit/fabrik/tui"
-	"github.com/handarbeit/fabrik/warnings"
 )
 
 func TestPoll_FetchesBoardAndProcessesItems(t *testing.T) {
@@ -1122,7 +1120,6 @@ func TestCruiseCatchUp_BothCruiseAndYolo_CruiseWins(t *testing.T) {
 	}
 }
 
-
 // TestPollPreFilter_WaitForCI_CompleteLabelOnly_Skipped verifies that an item with
 // wait_for_ci: true and ONLY stage:X:complete (no fabrik:awaiting-ci) is filtered by
 // the poll pre-filter when not in cycleSet and no expired CooldownAt. In the
@@ -1296,108 +1293,6 @@ func TestPoll_CruiseValidateComplete_NoRepeatDeepFetch(t *testing.T) {
 
 	if deepFetchCount > 1 {
 		t.Errorf("FetchItemDetails called %d times across 2 polls; want at most 1 — perpetual deep-fetch loop must not recur", deepFetchCount)
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Layer 1 — opportunistic per-event Status refresh
-// ---------------------------------------------------------------------------
-
-// seedTestCache creates a CacheImpl bootstrapped with one item (PVTI_001, owner/repo#1, "Research").
-func seedTestCache(t *testing.T, client *mockGitHubClient) *boardcache.CacheImpl {
-	t.Helper()
-	cache := boardcache.NewCacheImpl(client, itemstate.NewStore(nil), func(format string, args ...any) {})
-	board := &gh.ProjectBoard{
-		ProjectID: "PVT_test",
-		Items: []gh.ProjectItem{
-			{Number: 1, ItemID: "PVTI_001", Status: "Research", Repo: "owner/repo"},
-		},
-	}
-	testBootstrapFromBoard(cache, board)
-	return cache
-}
-
-func TestLayer1StatusRefresh_UpdatesCacheOnIssueCommentEvent(t *testing.T) {
-	client := &mockGitHubClient{
-		fetchProjectItemStatusFn: func(itemID string) (string, error) {
-			return "Plan", nil
-		},
-	}
-	cache := seedTestCache(t, client)
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	payload, _ := json.Marshal(map[string]any{
-		"issue":      map[string]any{"number": 1},
-		"repository": map[string]any{"full_name": "owner/repo"},
-	})
-	eng.applyLayer1StatusRefresh("issue_comment", payload, cache)
-
-	// FetchProjectItemStatus must have been called for PVTI_001.
-	client.mu.Lock()
-	calls := append([]string(nil), client.fetchProjectItemStatusCalls...)
-	client.mu.Unlock()
-	if len(calls) != 1 || calls[0] != "PVTI_001" {
-		t.Errorf("want FetchProjectItemStatus called once with %q, got %v", "PVTI_001", calls)
-	}
-
-	// Verify the cache now reflects "Plan" — the status returned by FetchProjectItemStatus.
-	board, err := cache.FetchProjectBoard("owner", "repo", 1, "")
-	if err != nil {
-		t.Fatalf("FetchProjectBoard: %v", err)
-	}
-	var gotStatus string
-	for _, item := range board.Items {
-		if item.Number == 1 {
-			gotStatus = item.Status
-		}
-	}
-	if gotStatus != "Plan" {
-		t.Errorf("want cached Status %q after Layer 1 refresh, got %q", "Plan", gotStatus)
-	}
-}
-
-func TestLayer1StatusRefresh_SkipsWhenCachePaused(t *testing.T) {
-	var callCount int32
-	client := &mockGitHubClient{
-		fetchProjectItemStatusFn: func(itemID string) (string, error) {
-			atomic.AddInt32(&callCount, 1)
-			return "Plan", nil
-		},
-	}
-	cache := seedTestCache(t, client)
-	cache.Pause()
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	payload, _ := json.Marshal(map[string]any{
-		"issue":      map[string]any{"number": 1},
-		"repository": map[string]any{"full_name": "owner/repo"},
-	})
-	eng.applyLayer1StatusRefresh("issue_comment", payload, cache)
-
-	if n := atomic.LoadInt32(&callCount); n != 0 {
-		t.Errorf("FetchProjectItemStatus should not be called when cache is paused; got %d calls", n)
-	}
-}
-
-func TestLayer1StatusRefresh_SkipsNonIssueEvents(t *testing.T) {
-	var callCount int32
-	client := &mockGitHubClient{
-		fetchProjectItemStatusFn: func(itemID string) (string, error) {
-			atomic.AddInt32(&callCount, 1)
-			return "Plan", nil
-		},
-	}
-	cache := seedTestCache(t, client)
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	payload, _ := json.Marshal(map[string]any{
-		"pull_request": map[string]any{"number": 1},
-		"repository":   map[string]any{"full_name": "owner/repo"},
-	})
-	eng.applyLayer1StatusRefresh("pull_request", payload, cache)
-
-	if n := atomic.LoadInt32(&callCount); n != 0 {
-		t.Errorf("FetchProjectItemStatus should not be called for pull_request events; got %d calls", n)
 	}
 }
 
@@ -2019,468 +1914,6 @@ func TestWorkerExitedIdempotent(t *testing.T) {
 	}
 }
 
-// ---------------------------------------------------------------------------
-// runProbeAndDeepFetch integration tests
-// ---------------------------------------------------------------------------
-
-// TestRunProbeAndDeepFetch_StaleItem_TriggersDeepFetch verifies that an item
-// with no prior deep-fetch (LastSeenSourceUpdatedAt == zero) triggers
-// FetchItemDetails when the probe returns a nonzero EffectiveUpdatedAt.
-func TestRunProbeAndDeepFetch_StaleItem_TriggersDeepFetch(t *testing.T) {
-	now := time.Now().Truncate(time.Second)
-	var deepFetchCalls int
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			return []gh.BoardProbeItem{
-				{ItemID: "PVTI_001", ContentID: "I_001", Number: 1, Repo: "owner/repo", Status: "Research", EffectiveUpdatedAt: now},
-			}, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error {
-			deepFetchCalls++
-			return nil
-		},
-	}
-	eng, cache := testEngineWithCache(t, client, &mockClaudeInvoker{})
-	eng.runProbeAndDeepFetch(cache)
-	if deepFetchCalls == 0 {
-		t.Error("expected FetchItemDetails called for stale item (zero LastSeenSourceUpdatedAt); got 0 calls")
-	}
-}
-
-// TestRunProbeAndDeepFetch_FreshItem_SkipsDeepFetch verifies that an item
-// whose LastSeenSourceUpdatedAt matches the probe's EffectiveUpdatedAt does
-// not trigger a FetchItemDetails call.
-func TestRunProbeAndDeepFetch_FreshItem_SkipsDeepFetch(t *testing.T) {
-	T1 := time.Now().Add(-time.Hour).Truncate(time.Second)
-	var deepFetchCalls int
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			return []gh.BoardProbeItem{
-				{ItemID: "PVTI_001", ContentID: "I_001", Number: 1, Repo: "owner/repo", Status: "Research", EffectiveUpdatedAt: T1},
-			}, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error {
-			deepFetchCalls++
-			return nil
-		},
-	}
-	eng, cache := testEngineWithCache(t, client, &mockClaudeInvoker{})
-	// Simulate a prior deep-fetch that set LastSeenSourceUpdatedAt = T1.
-	eng.store.Apply(itemstate.ItemDeepFetched{
-		Repo:   "owner/repo",
-		Number: 1,
-		FreshState: gh.ProjectItem{
-			ID: "I_001", ItemID: "PVTI_001", Number: 1, Repo: "owner/repo", UpdatedAt: T1,
-		},
-	})
-	eng.runProbeAndDeepFetch(cache)
-	if deepFetchCalls != 0 {
-		t.Errorf("expected 0 FetchItemDetails calls for fresh item; got %d", deepFetchCalls)
-	}
-}
-
-// TestRunProbeAndDeepFetch_LinkageDrift_InvalidatesAndDeepFetches verifies
-// that when the probe detects a linked PR number different from the cached
-// value, the cache is invalidated (DeepFetchInvalidated) and FetchItemDetails
-// is triggered even though EffectiveUpdatedAt has not advanced.
-func TestRunProbeAndDeepFetch_LinkageDrift_InvalidatesAndDeepFetches(t *testing.T) {
-	T1 := time.Now().Add(-time.Hour).Truncate(time.Second)
-	var deepFetchCalls int
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			return []gh.BoardProbeItem{
-				// Same EffectiveUpdatedAt as last deep-fetch (would be fresh) but LinkedPRNumber changed.
-				{ItemID: "PVTI_001", ContentID: "I_001", Number: 1, Repo: "owner/repo", Status: "Research", EffectiveUpdatedAt: T1, LinkedPRNumber: 99},
-			}, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error {
-			deepFetchCalls++
-			return nil
-		},
-	}
-	eng, cache := testEngineWithCache(t, client, &mockClaudeInvoker{})
-	// Simulate fresh state at T1 with no linked PR (cached LinkedPRNumber = 0).
-	eng.store.Apply(itemstate.ItemDeepFetched{
-		Repo:   "owner/repo",
-		Number: 1,
-		FreshState: gh.ProjectItem{
-			ID: "I_001", ItemID: "PVTI_001", Number: 1, Repo: "owner/repo", UpdatedAt: T1,
-		},
-	})
-	eng.runProbeAndDeepFetch(cache)
-	if deepFetchCalls == 0 {
-		t.Error("expected FetchItemDetails called after linkage drift (PR# 0 → 99); got 0 calls")
-	}
-}
-
-// TestRunProbeAndDeepFetch_ItemGone_RemovedFromStore verifies that an item
-// present in the store but absent from probe results is removed from the store.
-func TestRunProbeAndDeepFetch_ItemGone_RemovedFromStore(t *testing.T) {
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			// Only item #1; item #2 has left the board.
-			return []gh.BoardProbeItem{
-				{ItemID: "PVTI_001", ContentID: "I_001", Number: 1, Repo: "owner/repo", Status: "Research"},
-			}, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error { return nil },
-	}
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-	cache := boardcache.NewCacheImpl(client, eng.store, func(string, ...any) {})
-	testBootstrapFromBoard(cache, &gh.ProjectBoard{
-		ProjectID: "PVT_1",
-		Items: []gh.ProjectItem{
-			{ID: "I_001", ItemID: "PVTI_001", Number: 1, Repo: "owner/repo"},
-			{ID: "I_002", ItemID: "PVTI_002", Number: 2, Repo: "owner/repo"},
-		},
-	})
-	eng.readClient = cache
-
-	eng.runProbeAndDeepFetch(cache)
-
-	if _, err := eng.store.Get("owner/repo", 2); err == nil {
-		t.Error("item #2 should be removed from store after probe omits it")
-	}
-	if _, err := eng.store.Get("owner/repo", 1); err != nil {
-		t.Errorf("item #1 should still be in store after probe includes it: %v", err)
-	}
-}
-
-// TestRunStartupTransientLabelScan_RemovesStaleLabelsFromClosedItems verifies
-// that runStartupTransientLabelScan triggers label cleanup on closed store
-// entries carrying transient lifecycle labels, without touching open items or
-// clean closed items.
-func TestRunStartupTransientLabelScan_RemovesStaleLabelsFromClosedItems(t *testing.T) {
-	var removedLabels []string
-	client := &mockGitHubClient{
-		removeLabelFromIssueFn: func(owner, repo string, issueNumber int, labelName string) error {
-			removedLabels = append(removedLabels, labelName)
-			return nil
-		},
-	}
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	// Seed three items into the store:
-	//   #1 — closed, carries a transient label → should be cleaned
-	//   #2 — open, carries a transient label → must be skipped
-	//   #3 — closed, no transient labels → must be skipped
-	for _, pi := range []gh.ProjectItem{
-		{ID: "I_001", Number: 1, Repo: "owner/repo", IsClosed: true, Labels: []string{"fabrik:awaiting-review", "stage:Review:complete"}},
-		{ID: "I_002", Number: 2, Repo: "owner/repo", IsClosed: false, Labels: []string{"fabrik:awaiting-ci"}},
-		{ID: "I_003", Number: 3, Repo: "owner/repo", IsClosed: true, Labels: []string{"stage:Validate:complete"}},
-	} {
-		eng.store.Apply(itemstate.IssueOpened{Item: pi})
-		eng.store.Apply(itemstate.ItemDeepFetched{
-			Repo:       pi.Repo,
-			Number:     pi.Number,
-			FreshState: pi,
-		})
-	}
-
-	eng.runStartupTransientLabelScan()
-
-	// Only the transient label from issue #1 should be removed.
-	if len(removedLabels) == 0 {
-		t.Fatal("expected RemoveLabelFromIssue called for closed item with stale transient label; got 0 calls")
-	}
-	found := false
-	for _, l := range removedLabels {
-		if l == "fabrik:awaiting-review" {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("expected 'fabrik:awaiting-review' to be removed; removed labels: %v", removedLabels)
-	}
-}
-
-// TestColdStart_ProbeBootstrap_TerminalItemsSkipDeepFetch verifies the cold-start cost
-// reduction: 10 closed Done items are seeded terminal by BootstrapFromProbe and
-// are never deep-fetched, while 3 open active items are deep-fetched normally.
-// Expected deep-fetch count after the first probe cycle: ≤ 3.
-func TestColdStart_ProbeBootstrap_TerminalItemsSkipDeepFetch(t *testing.T) {
-	var deepFetchCalls int
-	probeTime := time.Now().Add(-time.Minute)
-
-	// Build 10 closed Done items + 3 open Research items for the probe response.
-	var probeItems []gh.BoardProbeItem
-	for i := 1; i <= 10; i++ {
-		probeItems = append(probeItems, gh.BoardProbeItem{
-			ContentID: fmt.Sprintf("I_%03d", i), ItemID: fmt.Sprintf("PVTI_%03d", i),
-			Number: i, Repo: "owner/repo",
-			Status:             "Done",
-			IsClosed:           true,
-			EffectiveUpdatedAt: probeTime,
-		})
-	}
-	for i := 11; i <= 13; i++ {
-		probeItems = append(probeItems, gh.BoardProbeItem{
-			ContentID: fmt.Sprintf("I_%03d", i), ItemID: fmt.Sprintf("PVTI_%03d", i),
-			Number: i, Repo: "owner/repo",
-			Status:             "Research",
-			IsClosed:           false,
-			EffectiveUpdatedAt: probeTime,
-		})
-	}
-
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			return probeItems, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error {
-			deepFetchCalls++
-			return nil
-		},
-	}
-	eng := NewWithDeps(
-		Config{
-			Owner: "owner", Repo: "repo", ProjectNum: 1,
-			User: "testuser", Token: "token", MaxConcurrent: 5,
-			Stages: testStagesWithCleanup(),
-		},
-		client, &mockClaudeInvoker{}, NewWorktreeManager(t.TempDir()),
-	)
-	cache := boardcache.NewCacheImpl(client, eng.store, func(string, ...any) {})
-	eng.readClient = cache
-
-	// Simulate the virgin-cache branch: probe bootstrap seeds the store.
-	items, projectID, err := client.ProbeProjectBoard("owner", "repo", 1, "organization")
-	if err != nil {
-		t.Fatalf("ProbeProjectBoard: %v", err)
-	}
-	cache.BootstrapFromProbe(items, projectID)
-	eng.seedTerminalFromProbeItems(items)
-
-	// Now simulate the next poll cycle — probe-driven deep-fetch pass.
-	eng.runProbeAndDeepFetch(cache)
-
-	// The 10 closed Done items are seeded terminal and must NOT be deep-fetched.
-	// The 3 open Research items have no prior deep-fetch and MUST be deep-fetched.
-	if deepFetchCalls > 3 {
-		t.Errorf("cold-start deep-fetch count = %d, want ≤ 3 (only active items)", deepFetchCalls)
-	}
-	if deepFetchCalls == 0 {
-		t.Error("expected ≥ 1 deep-fetch for active items; got 0")
-	}
-
-	// Terminal flag must be set on all closed Done items.
-	for i := 1; i <= 10; i++ {
-		snap, snapErr := eng.store.Get("owner/repo", i)
-		if snapErr != nil {
-			t.Errorf("item #%d not found in store", i)
-			continue
-		}
-		if !snap.IsTerminal() {
-			t.Errorf("item #%d (closed Done): expected IsTerminal()=true", i)
-		}
-	}
-
-	// Active Research items must NOT be terminal.
-	for i := 11; i <= 13; i++ {
-		snap, snapErr := eng.store.Get("owner/repo", i)
-		if snapErr != nil {
-			t.Errorf("item #%d not found in store", i)
-			continue
-		}
-		if snap.IsTerminal() {
-			t.Errorf("item #%d (open Research): expected IsTerminal()=false", i)
-		}
-	}
-}
-
-// TestWebhookModeStartup_ClosedDoneItemsNotDeepFetched is the regression test
-// for issue #751. It verifies that after the fixed webhook-mode startup path
-// (BootstrapFromProbe instead of Bootstrap), the first probe cycle does NOT
-// call FetchItemDetails for closed Done items. Only active items are fetched.
-func TestWebhookModeStartup_ClosedDoneItemsNotDeepFetched(t *testing.T) {
-	var deepFetchCalls int
-	probeTime := time.Now().Add(-time.Minute)
-
-	// 3 closed Done items + 1 open Research item returned by the probe.
-	allProbeItems := []gh.BoardProbeItem{
-		{ContentID: "I_001", ItemID: "PVTI_001", Number: 1, Repo: "owner/repo",
-			Status: "Done", IsClosed: true, EffectiveUpdatedAt: probeTime},
-		{ContentID: "I_002", ItemID: "PVTI_002", Number: 2, Repo: "owner/repo",
-			Status: "Done", IsClosed: true, EffectiveUpdatedAt: probeTime},
-		{ContentID: "I_003", ItemID: "PVTI_003", Number: 3, Repo: "owner/repo",
-			Status: "Done", IsClosed: true, EffectiveUpdatedAt: probeTime},
-		{ContentID: "I_004", ItemID: "PVTI_004", Number: 4, Repo: "owner/repo",
-			Status: "Research", IsClosed: false, EffectiveUpdatedAt: probeTime},
-	}
-
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			return allProbeItems, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error {
-			deepFetchCalls++
-			return nil
-		},
-	}
-	eng := NewWithDeps(
-		Config{
-			Owner: "owner", Repo: "repo", ProjectNum: 1,
-			User: "testuser", Token: "token", MaxConcurrent: 5,
-			Stages: testStagesWithCleanup(),
-		},
-		client, &mockClaudeInvoker{}, NewWorktreeManager(t.TempDir()),
-	)
-	cache := boardcache.NewCacheImpl(client, eng.store, func(string, ...any) {})
-
-	// Simulate the fixed webhook startup path: BootstrapFromProbe seeds Terminal
-	// for closed Done items before the first poll cycle runs.
-	cache.BootstrapFromProbe(allProbeItems, "PVT_1")
-	eng.seedTerminalFromProbeItems(allProbeItems)
-	eng.readClient = cache
-
-	// Run one probe cycle — this is what the first poll does after startup.
-	eng.runProbeAndDeepFetch(cache)
-
-	// The 3 closed Done items are seeded terminal and must NOT be deep-fetched.
-	// Only the 1 open Research item should trigger FetchItemDetails.
-	if deepFetchCalls != 1 {
-		t.Errorf("webhook startup deep-fetch count = %d, want 1 (only active Research item)", deepFetchCalls)
-	}
-
-	// Verify terminal flag is set for closed Done items.
-	for i := 1; i <= 3; i++ {
-		snap, err := eng.store.Get("owner/repo", i)
-		if err != nil {
-			t.Errorf("item #%d not in store: %v", i, err)
-			continue
-		}
-		if !snap.IsTerminal() {
-			t.Errorf("item #%d (closed Done): expected IsTerminal()=true after webhook startup", i)
-		}
-	}
-}
-
-// TestRunProbeAndDeepFetch_IsClosedPropagates_WithoutDeepFetch verifies that
-// IsClosed=true is written to the store via ProbeBoardItemUpdated even when
-// the item is cache-fresh (EffectiveUpdatedAt unchanged → no deep-fetch).
-func TestRunProbeAndDeepFetch_IsClosedPropagates_WithoutDeepFetch(t *testing.T) {
-	T1 := time.Now().Add(-time.Hour).Truncate(time.Second)
-	var deepFetchCalls int
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			return []gh.BoardProbeItem{
-				{ItemID: "PVTI_001", ContentID: "I_001", Number: 1, Repo: "owner/repo", Status: "Research", IsClosed: true, EffectiveUpdatedAt: T1},
-			}, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error {
-			deepFetchCalls++
-			return nil
-		},
-	}
-	eng, cache := testEngineWithCache(t, client, &mockClaudeInvoker{})
-	// Fresh at T1 — deep-fetch should not be triggered.
-	eng.store.Apply(itemstate.ItemDeepFetched{
-		Repo:   "owner/repo",
-		Number: 1,
-		FreshState: gh.ProjectItem{
-			ID: "I_001", ItemID: "PVTI_001", Number: 1, Repo: "owner/repo", UpdatedAt: T1,
-		},
-	})
-	eng.runProbeAndDeepFetch(cache)
-	if deepFetchCalls != 0 {
-		t.Errorf("expected 0 deep-fetch calls for fresh item; got %d", deepFetchCalls)
-	}
-	snap, err := eng.store.Get("owner/repo", 1)
-	if err != nil {
-		t.Fatalf("store.Get after probe: %v", err)
-	}
-	if !snap.IsClosed() {
-		t.Error("expected IsClosed=true after ProbeBoardItemUpdated; got false")
-	}
-}
-
-// TestProbeNewItem_ClosedDone_SkipsDeepFetch is a regression test for the
-// new-item branch of runProbeAndDeepFetch. It verifies that closed Done items
-// discovered by the probe (not yet in store, no prior bootstrap) are seeded as
-// terminal and never deep-fetched, while open active items are deep-fetched
-// normally. This covers the gap where BootstrapFromProbe cannot help: items
-// that appear in the probe for the first time during a mid-run cycle.
-func TestProbeNewItem_ClosedDone_SkipsDeepFetch(t *testing.T) {
-	const numClosed = 3 // closed Done items
-	const numOpen = 2   // open Research items
-	var deepFetchCalls int
-	probeTime := time.Now().Add(-time.Minute)
-
-	var probeItems []gh.BoardProbeItem
-	for i := 1; i <= numClosed; i++ {
-		probeItems = append(probeItems, gh.BoardProbeItem{
-			ContentID: fmt.Sprintf("I_%03d", i), ItemID: fmt.Sprintf("PVTI_%03d", i),
-			Number:             i,
-			Repo:               "owner/repo",
-			Status:             "Done",
-			IsClosed:           true,
-			EffectiveUpdatedAt: probeTime,
-		})
-	}
-	for i := numClosed + 1; i <= numClosed+numOpen; i++ {
-		probeItems = append(probeItems, gh.BoardProbeItem{
-			ContentID: fmt.Sprintf("I_%03d", i), ItemID: fmt.Sprintf("PVTI_%03d", i),
-			Number:             i,
-			Repo:               "owner/repo",
-			Status:             "Research",
-			IsClosed:           false,
-			EffectiveUpdatedAt: probeTime,
-		})
-	}
-
-	client := &mockGitHubClient{
-		probeProjectBoardFn: func(owner, repo string, projectNum int, ownerType string) ([]gh.BoardProbeItem, string, error) {
-			return probeItems, "PVT_1", nil
-		},
-		fetchItemDetailsFn: func(item *gh.ProjectItem) error {
-			deepFetchCalls++
-			return nil
-		},
-	}
-	eng := NewWithDeps(
-		Config{
-			Owner: "owner", Repo: "repo", ProjectNum: 1,
-			User: "testuser", Token: "token", MaxConcurrent: 5,
-			Stages: testStagesWithCleanup(),
-		},
-		client, &mockClaudeInvoker{}, NewWorktreeManager(t.TempDir()),
-	)
-	cache := boardcache.NewCacheImpl(client, eng.store, func(string, ...any) {})
-	eng.readClient = cache
-
-	// No prior bootstrap — store is empty. All items are new-item discoveries.
-	eng.runProbeAndDeepFetch(cache)
-
-	// Only the open Research items should have been deep-fetched.
-	if deepFetchCalls != numOpen {
-		t.Errorf("deep-fetch count = %d, want %d (open items only)", deepFetchCalls, numOpen)
-	}
-
-	// Closed Done items must be terminal in the store.
-	for i := 1; i <= numClosed; i++ {
-		snap, snapErr := eng.store.Get("owner/repo", i)
-		if snapErr != nil {
-			t.Errorf("closed Done item #%d not found in store", i)
-			continue
-		}
-		if !snap.IsTerminal() {
-			t.Errorf("item #%d (closed Done): expected IsTerminal()=true", i)
-		}
-	}
-
-	// Open Research items must NOT be terminal.
-	for i := numClosed + 1; i <= numClosed+numOpen; i++ {
-		snap, snapErr := eng.store.Get("owner/repo", i)
-		if snapErr != nil {
-			t.Errorf("open Research item #%d not found in store", i)
-			continue
-		}
-		if snap.IsTerminal() {
-			t.Errorf("item #%d (open Research): expected IsTerminal()=false", i)
-		}
-	}
-}
-
 // captureStdout redirects os.Stdout to a pipe for the duration of fn,
 // then returns everything written to stdout. The original os.Stdout is
 // restored before returning.
@@ -2494,103 +1927,6 @@ func captureStdout(fn func()) string {
 	var buf strings.Builder
 	io.Copy(&buf, r)
 	return buf.String()
-}
-
-func TestCheckAllowAutoMerge_DisabledEmitsWarning(t *testing.T) {
-	warnings.WarningsPathOverride = filepath.Join(t.TempDir(), "warnings.json")
-	t.Cleanup(func() { warnings.WarningsPathOverride = "" })
-	client := &mockGitHubClient{
-		fetchAllowAutoMergeFn: func(owner, repo string) (bool, error) {
-			return false, nil
-		},
-	}
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	out := captureStdout(func() {
-		eng.checkAllowAutoMerge("owner", "repo")
-	})
-
-	if !strings.Contains(out, "WARNING") {
-		t.Errorf("expected WARNING in output; got: %q", out)
-	}
-	if !strings.Contains(out, "allow_auto_merge") {
-		t.Errorf("expected allow_auto_merge mention in output; got: %q", out)
-	}
-	if !strings.Contains(out, "gh api -X PATCH repos/owner/repo") {
-		t.Errorf("expected fix command in output; got: %q", out)
-	}
-}
-
-func TestCheckAllowAutoMerge_EnabledIsSilent(t *testing.T) {
-	warnings.WarningsPathOverride = filepath.Join(t.TempDir(), "warnings.json")
-	t.Cleanup(func() { warnings.WarningsPathOverride = "" })
-	client := &mockGitHubClient{
-		fetchAllowAutoMergeFn: func(owner, repo string) (bool, error) {
-			return true, nil
-		},
-	}
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	out := captureStdout(func() {
-		eng.checkAllowAutoMerge("owner", "repo")
-	})
-
-	if out != "" {
-		t.Errorf("expected no output for enabled repo; got: %q", out)
-	}
-}
-
-func TestCheckAllowAutoMerge_APIErrorIsNonFatal(t *testing.T) {
-	warnings.WarningsPathOverride = filepath.Join(t.TempDir(), "warnings.json")
-	t.Cleanup(func() { warnings.WarningsPathOverride = "" })
-	client := &mockGitHubClient{
-		fetchAllowAutoMergeFn: func(owner, repo string) (bool, error) {
-			return false, errors.New("network error")
-		},
-	}
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	// Should not panic; engine should log the error at warn level and continue.
-	out := captureStdout(func() {
-		eng.checkAllowAutoMerge("owner", "repo")
-	})
-
-	// No WARNING block should be emitted for an API error.
-	if strings.Contains(out, "WARNING") {
-		t.Errorf("should not print WARNING on API error; got: %q", out)
-	}
-}
-
-func TestCheckAllowAutoMerge_DedupSuppressesSecondCall(t *testing.T) {
-	warnings.WarningsPathOverride = filepath.Join(t.TempDir(), "warnings.json")
-	t.Cleanup(func() { warnings.WarningsPathOverride = "" })
-	var callCount int
-	client := &mockGitHubClient{
-		fetchAllowAutoMergeFn: func(owner, repo string) (bool, error) {
-			callCount++
-			return false, nil
-		},
-	}
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-
-	// First call should emit warning.
-	out1 := captureStdout(func() {
-		eng.checkAllowAutoMerge("owner", "repo")
-	})
-	if !strings.Contains(out1, "WARNING") {
-		t.Errorf("first call: expected WARNING; got: %q", out1)
-	}
-
-	// Second call for the same repo should be a no-op.
-	out2 := captureStdout(func() {
-		eng.checkAllowAutoMerge("owner", "repo")
-	})
-	if out2 != "" {
-		t.Errorf("second call: expected no output (dedup); got: %q", out2)
-	}
-	if callCount != 1 {
-		t.Errorf("expected API to be called exactly once; got %d", callCount)
-	}
 }
 
 // TestPoll_InFlightWorker_NotSupplanted is a regression test for the
@@ -3027,9 +2363,9 @@ func TestHandleMergeTrainBatch_LogsQueuedItems(t *testing.T) {
 	board := &gh.ProjectBoard{
 		ProjectID: "PVT_1",
 		Items: []gh.ProjectItem{
-			{Number: 42, Title: "fix the bug",    Status: "BatchHold"},
+			{Number: 42, Title: "fix the bug", Status: "BatchHold"},
 			{Number: 43, Title: "add the feature", Status: "Implement"},
-			{Number: 44, Title: "another ready",   Status: "BatchHold"},
+			{Number: 44, Title: "another ready", Status: "BatchHold"},
 		},
 	}
 
@@ -3076,7 +2412,7 @@ func TestHandleMergeTrainBatch_SilentWhenEmpty(t *testing.T) {
 		ProjectID: "PVT_1",
 		Items: []gh.ProjectItem{
 			{Number: 10, Title: "in progress", Status: "Implement"},
-			{Number: 11, Title: "reviewing",   Status: "Review"},
+			{Number: 11, Title: "reviewing", Status: "Review"},
 		},
 	}
 
@@ -3390,64 +2726,5 @@ func TestGroupQueuedByRepo_ExcludesPausedAndClosed(t *testing.T) {
 	}
 	if len(groups[0].items) != 2 || groups[0].items[0].Number != 1 || groups[0].items[1].Number != 4 {
 		t.Errorf("expected only clean members #1 and #4, got %+v", groups[0].items)
-	}
-}
-
-// TestReconcileLoop_RunsWithoutWebhookManager is the #955 regression for the
-// architectural fix: the reconcile ticker must run — and repair label drift — even
-// when the webhook manager is nil (webhooks disabled or wm.Start failed).
-// Previously the ticker was nested inside the webhook-start block, so a webhook-less
-// deployment never reconciled and could not self-heal a drifted fabrik label set,
-// stranding items at gates (e.g. fabrik:awaiting-ci missing from the store forever).
-func TestReconcileLoop_RunsWithoutWebhookManager(t *testing.T) {
-	t1 := time.Now().Truncate(time.Second)
-	client := &mockGitHubClient{}
-	eng := testEngine(t, client, &mockClaudeInvoker{})
-	eng.cfg.ReconcileInterval = 10 * time.Millisecond
-
-	cache := boardcache.NewCacheImpl(client, eng.store, func(string, ...any) {})
-	// Seed #1 at Validate WITHOUT fabrik:awaiting-ci — the drifted store state.
-	testBootstrapFromBoard(cache, &gh.ProjectBoard{
-		ProjectID: "PVT_1",
-		Items: []gh.ProjectItem{
-			{ID: "I_1", ItemID: "PVTI_1", Number: 1, Repo: "owner/repo", Status: "Validate", Labels: []string{"fabrik:cruise"}, UpdatedAt: t1},
-		},
-	})
-	eng.readClient = cache
-
-	// GitHub's fresh board: same status + updatedAt, but WITH fabrik:awaiting-ci.
-	// Only the fabrik-managed label differs — exactly the #1479 stranding condition.
-	client.fetchProjectBoardFn = func(_, _ string, _ int, _ string) (*gh.ProjectBoard, error) {
-		return &gh.ProjectBoard{
-			ProjectID: "PVT_1",
-			Items: []gh.ProjectItem{
-				{ID: "I_1", ItemID: "PVTI_1", Number: 1, Repo: "owner/repo", Status: "Validate", Labels: []string{"fabrik:cruise", "fabrik:awaiting-ci"}, UpdatedAt: t1},
-			},
-		}, nil
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	// nil webhook manager — the whole point: reconcile must run anyway.
-	go eng.reconcileLoop(ctx, cache, nil)
-
-	deadline := time.Now().Add(2 * time.Second)
-	for {
-		snap, err := eng.store.Get("owner/repo", 1)
-		if err == nil {
-			for _, l := range snap.Labels() {
-				if l == "fabrik:awaiting-ci" {
-					return // success: reconcile ran with nil wm and synced the gate label
-				}
-			}
-		}
-		if time.Now().After(deadline) {
-			var labels []string
-			if err == nil {
-				labels = snap.Labels()
-			}
-			t.Fatalf("reconcileLoop did not sync fabrik:awaiting-ci within 2s (nil wm); labels = %v", labels)
-		}
-		time.Sleep(10 * time.Millisecond)
 	}
 }
